+++ run_on_coal/Scene/Camera.h
@@ -11,8 +11,7 @@ class Camera
     bool m_rebuildViewMatrix;
     void UpdateViewMatrix();
 public:
-
-    void SetPerspective(float f_fov,int f_width, int f_height, float f_near, float f_far);
+    void SetPerspective(float f_fov, float f_width, float f_height, float f_near, float f_far);
     void SetOrtho(float l_left,float l_right,float f_bottom,float f_top,float l_near,float l_far);
 
     void SetPosition(glm::vec3 &f_pos);
